[NOTIFICATION]: refactor config package
