+++ notification/src/main/java/io/skai/notification/config/email/GmailProperties.java
@@ -1,4 +1,4 @@
-package io.skai.notification.config;
+package io.skai.notification.config.email;
 
 import lombok.Getter;
 import lombok.Setter;
